+++ app/models.py
@@ -41,8 +41,8 @@ class Permission(db.Model):
         cascade='all, delete-orphan'
     )
 
-    def roles_alias(self, category=None):
-        '''Permission.roles_alias(self, category=None)'''
+    def roles_from_category(self, category=None):
+        '''Permission.roles_from_category(self, category=None)'''
         if category is not None:
             roles = Role.query\
                 .join(RolePermission, RolePermission.role_id == Role.id)\
@@ -106,24 +106,12 @@ class Role(db.Model):
     )
     users = db.relationship('User', backref='role', lazy='dynamic')
 
-    def add_permission(self, permission):
-        '''Role.add_permission(self, permission)'''
-        if not self.has_permission(permission=permission):
-            role_permission = RolePermission(role_id=self.id, permission_id=permission.id)
-            db.session.add(role_permission)
-
-    def remove_permission(self, permission):
-        '''Role.remove_permission(self, permission)'''
-        role_permission = self.permissions.filter_by(permission_id=permission.id).first()
-        if role_permission is not None:
-            db.session.delete(role_permission)
-
     def has_permission(self, permission):
         '''Role.has_permission(self, permission)'''
         return self.permissions.filter_by(permission_id=permission.id).first() is not None
 
-    def permissions_alias(self, category=None):
-        '''Role.permissions_alias(self, category=None)'''
+    def permissions_from_category(self, category=None):
+        '''Role.permissions_from_category(self, category=None)'''
         if category is not None:
             permissions = Permission.query\
                 .join(RolePermission, RolePermission.permission_id == Permission.id)\
@@ -179,7 +167,11 @@ class Role(db.Model):
                 if verbose:
                     print('导入用户角色信息', entry['name'], entry['category'])
                 for permission in entry['permissions']:
-                    role.add_permission(permission=Permission.query.filter_by(name=permission).first())
+                    role_permission = RolePermission(
+                        role_id=role.id,
+                        permission_id=Permission.query.filter_by(name=permission).first().id
+                    )
+                    db.session.add(role_permission)
                     if verbose:
                         print('赋予权限', entry['name'], permission)
                 db.session.commit()
@@ -659,6 +651,7 @@ class User(UserMixin, db.Model):
                 Lesson.order > 0,
                 Lesson.order <= lesson.order
             ))\
+            .filter(Video.restricted == False)\
             .all():
             self.punch(video=video, play_time=video.duration)
 
@@ -680,6 +673,7 @@ class User(UserMixin, db.Model):
             .join(LessonType, LessonType.id == Lesson.type_id)\
             .filter(LessonType.name == 'VB')\
             .filter(Lesson.order > 0)\
+            .filter(Video.restricted == False)\
             .filter(Punch.user_id == self.id)\
             .order_by(Video.id.desc())\
             .first()
@@ -693,6 +687,7 @@ class User(UserMixin, db.Model):
             .join(LessonType, LessonType.id == Lesson.type_id)\
             .filter(LessonType.name == 'Y-GRE')\
             .filter(Lesson.order > 0)\
+            .filter(Video.restricted == False)\
             .filter(Punch.user_id == self.id)\
             .order_by(Video.id.desc())\
             .first()
@@ -705,12 +700,14 @@ class User(UserMixin, db.Model):
                 .join(Lesson, Lesson.id == Video.lesson_id)\
                 .join(LessonType, LessonType.id == Lesson.type_id)\
                 .filter(LessonType.name == 'VB')\
+                .filter(Video.restricted == False)\
                 .all()], timedelta()).total_seconds(),
             'value': reduce(operator.add, [punch.play_time_trim for punch in Punch.query\
                 .join(Video, Video.id == Punch.video_id)\
                 .join(Lesson, Lesson.id == Video.lesson_id)\
                 .join(LessonType, LessonType.id == Lesson.type_id)\
                 .filter(LessonType.name == 'VB')\
+                .filter(Video.restricted == False)\
                 .filter(Punch.user_id == self.id)\
                 .all()], timedelta()).total_seconds(),
         }
@@ -728,12 +725,14 @@ class User(UserMixin, db.Model):
                 .join(Lesson, Lesson.id == Video.lesson_id)\
                 .join(LessonType, LessonType.id == Lesson.type_id)\
                 .filter(LessonType.name == 'Y-GRE')\
+                .filter(Video.restricted == False)\
                 .all()], timedelta()).total_seconds(),
             'value': reduce(operator.add, [punch.play_time_trim for punch in Punch.query\
                 .join(Video, Video.id == Punch.video_id)\
                 .join(Lesson, Lesson.id == Video.lesson_id)\
                 .join(LessonType, LessonType.id == Lesson.type_id)\
                 .filter(LessonType.name == 'Y-GRE')\
+                .filter(Video.restricted == False)\
                 .filter(Punch.user_id == self.id)\
                 .all()], timedelta()).total_seconds(),
         }
@@ -748,6 +747,7 @@ class User(UserMixin, db.Model):
         return reduce(operator.add, [punch.play_time_trim for punch in Punch.query\
             .join(Video, Video.id == Punch.video_id)\
             .filter(Video.lesson_id == lesson.id)\
+            .filter(Video.restricted == False)\
             .filter(Punch.user_id == self.id)\
             .all()], timedelta()) / lesson.duration
 
@@ -760,6 +760,7 @@ class User(UserMixin, db.Model):
         return Punch.query\
             .join(Video, Video.id == Punch.video_id)\
             .filter(Video.lesson_id == lesson.id)\
+            .filter(Video.restricted == False)\
             .filter(Punch.user_id == self.id)\
             .order_by(Punch.timestamp.desc())\
             .first()
@@ -801,6 +802,7 @@ class User(UserMixin, db.Model):
                 Lesson.order > 0,
                 Lesson.order < lesson.order
             ))\
+            .filter(Video.restricted == False)\
             .all()], True)
 
     def can_play(self, video):
@@ -1055,6 +1057,7 @@ class Device(db.Model):
     type_id = db.Column(db.Integer, db.ForeignKey('device_types.id'))
     room_id = db.Column(db.Integer, db.ForeignKey('rooms.id'))
     mac_address = db.Column(db.Unicode(64))
+    restricted_permit = db.Column(db.Boolean, default=False)
     category = db.Column(db.Unicode(64), default='production', index=True)
     obsolete = db.Column(db.Boolean, default=False)
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
@@ -1091,6 +1094,7 @@ class Device(db.Model):
             room,
             self.mac_address,
             self.category,
+            str(int(self.restricted_permit)),
             str(int(self.obsolete)),
             self.created_at.strftime(current_app.config['DATETIME_FORMAT']),
             self.modified_at.strftime(current_app.config['DATETIME_FORMAT']),
@@ -1108,6 +1112,7 @@ class Device(db.Model):
                 alias='Y-VOD Server',
                 type_id=DeviceType.query.filter_by(name='Server').first().id,
                 mac_address=current_app.config['SERVER_MAC_ADDRESS'],
+                restricted_permit=True,
                 category='development',
                 modified_by_id=User.query.get(1).id
             )
@@ -1131,12 +1136,13 @@ class Device(db.Model):
                                 type_id=DeviceType.query.filter_by(name=entry[2]).first().id,
                                 room_id=entry[3],
                                 mac_address=entry[4],
-                                category=entry[5],
+                                restricted_permit=bool(int(entry[5])),
+                                category=entry[6],
                                 modified_by_id=User.query.get(1).id
                             )
                             db.session.add(device)
                             if verbose:
-                                print('导入设备信息', entry[0], entry[1], entry[2], entry[4], entry[5])
+                                print('导入设备信息', entry[0], entry[1], entry[2], entry[4], entry[6])
                         else:
                             if entry[4] is not None:
                                 entry[4] = Room.query.filter_by(name=entry[4]).first().id
@@ -1147,15 +1153,16 @@ class Device(db.Model):
                                 type_id=DeviceType.query.filter_by(name=entry[3]).first().id,
                                 room_id=entry[4],
                                 mac_address=entry[5],
-                                category=entry[6],
-                                obsolete=bool(int(entry[7])),
-                                created_at=datetime.strptime(entry[8], current_app.config['DATETIME_FORMAT']),
-                                modified_at=datetime.strptime(entry[9], current_app.config['DATETIME_FORMAT']),
-                                modified_by_id=int(entry[10])
+                                restricted_permit=bool(int(entry[6])),
+                                category=entry[7],
+                                obsolete=bool(int(entry[8])),
+                                created_at=datetime.strptime(entry[9], current_app.config['DATETIME_FORMAT']),
+                                modified_at=datetime.strptime(entry[10], current_app.config['DATETIME_FORMAT']),
+                                modified_by_id=int(entry[11])
                             )
                             db.session.add(device)
                             if verbose:
-                                print('导入设备信息', entry[2], entry[1], entry[3], entry[5], entry[6])
+                                print('导入设备信息', entry[2], entry[1], entry[3], entry[5], entry[7])
                     line_num += 1
                 db.session.commit()
         else:
@@ -1176,6 +1183,7 @@ class Device(db.Model):
                 'type',
                 'room',
                 'mac_address',
+                'restricted_permit',
                 'category',
                 'obsolete',
                 'created_at',
@@ -1199,8 +1207,8 @@ class LessonType(db.Model):
     lessons = db.relationship('Lesson', backref='type', lazy='dynamic')
 
     @property
-    def snake_case(self):
-        '''LessonType.snake_case(self)'''
+    def name_snake_case(self):
+        '''LessonType.name_snake_case(self)'''
         return self.name.lower().replace('-', '_')
 
     @staticmethod
@@ -1237,9 +1245,19 @@ class Lesson(db.Model):
     videos = db.relationship('Video', backref='lesson', lazy='dynamic')
 
     @property
+    def unrestricted_videos(self):
+        '''Lesson.unrestricted_videos(self)'''
+        return self.videos.filter_by(restricted=False)
+
+    @property
+    def restricted_videos(self):
+        '''Lesson.restricted_videos(self)'''
+        return self.videos.filter_by(restricted=True)
+
+    @property
     def duration(self):
         '''Lesson.duration(self)'''
-        return reduce(operator.add, [video.duration for video in self.videos], timedelta())
+        return reduce(operator.add, [video.duration for video in self.unrestricted_videos], timedelta())
 
     @property
     def duration_format(self):
@@ -1298,6 +1316,50 @@ class Lesson(db.Model):
         return '<Lesson {}>'.format(self.name)
 
 
+class VideoCollection(db.Model):
+    '''Table: video_collections'''
+    __tablename__ = 'video_collections'
+    collection_id = db.Column(db.Integer, db.ForeignKey('collections.id'), primary_key=True)
+    video_id = db.Column(db.Integer, db.ForeignKey('videos.id'), primary_key=True)
+
+
+class Collection(db.Model):
+    '''Table: collections'''
+    __tablename__ = 'collections'
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.Unicode(64), unique=True, index=True)
+    videos = db.relationship(
+        'VideoCollection',
+        foreign_keys=[VideoCollection.collection_id],
+        backref=db.backref('collection', lazy='joined'),
+        lazy='dynamic',
+        cascade='all, delete-orphan'
+    )
+
+    def has_video(self, video):
+        '''Collection.has_video(self, video)'''
+        return self.videos.filter_by(video_id=video.id).first() is not None
+
+    @staticmethod
+    def insert_entries(data, verbose=False):
+        '''Collection.insert_entries(data, verbose=False)'''
+        yaml_file = os.path.join(current_app.config['DATA_DIR'], data, 'collections.yml')
+        entries = load_yaml(yaml_file=yaml_file)
+        if entries is not None:
+            print('---> Read: {}'.format(yaml_file))
+            for entry in entries:
+                collection = Collection(name=entry['name'])
+                db.session.add(collection)
+                if verbose:
+                    print('导入视频选集信息', entry['name'])
+            db.session.commit()
+        else:
+            print('文件不存在', yaml_file)
+
+    def __repr__(self):
+        return '<Collection {}>'.format(self.name)
+
+
 class Video(db.Model):
     '''Table: videos'''
     __tablename__ = 'videos'
@@ -1307,10 +1369,17 @@ class Video(db.Model):
     description = db.Column(db.Unicode(64))
     lesson_id = db.Column(db.Integer, db.ForeignKey('lessons.id'))
     duration = db.Column(db.Interval, default=timedelta())
-    demo = db.Column(db.Boolean, default=False)
     file_name = db.Column(db.Unicode(64))
     hls_cache_file_name = db.Column(db.Unicode(64))
+    restricted = db.Column(db.Boolean, default=False)
     timestamp = db.Column(db.DateTime, default=datetime.utcnow)
+    collections = db.relationship(
+        'VideoCollection',
+        foreign_keys=[VideoCollection.video_id],
+        backref=db.backref('video', lazy='joined'),
+        lazy='dynamic',
+        cascade='all, delete-orphan'
+    )
     punches = db.relationship(
         'Punch',
         foreign_keys=[Punch.video_id],
@@ -1318,6 +1387,13 @@ class Video(db.Model):
         lazy='dynamic',
         cascade='all, delete-orphan'
     )
+
+    @property
+    def collections_format(self):
+        '''Video.collections_format(self)'''
+        if self.collections.count():
+            return '、'.join([item.collection.name for item in self.collections])
+        return 'N/A'
 
     @property
     def duration_format(self):
@@ -1385,13 +1461,22 @@ class Video(db.Model):
                         description=entry['description'],
                         lesson_id=Lesson.query.filter_by(name=entry['lesson_name']).first().id,
                         duration=get_video_duration(video_file),
-                        demo=entry['demo'],
                         file_name=entry['file_name'],
-                        hls_cache_file_name=hls_cache_file_name
+                        hls_cache_file_name=hls_cache_file_name,
+                        restricted=entry['restricted']
                     )
                     db.session.add(video)
+                    db.session.commit()
                     if verbose:
                         print('导入视频信息', entry['lesson_name'], entry['abbr'], entry['file_name'])
+                    for collection in entry['collections']:
+                        video_collection = VideoCollection(
+                            collection_id=Collection.query.filter_by(name=collection).first().id,
+                            video_id=video.id
+                        )
+                        db.session.add(video_collection)
+                        if verbose:
+                            print('加入选集', collection, video.name)
                 else:
                     print('视频文件不存在', entry['lesson_name'], entry['abbr'], entry['file_name'])
             db.session.commit()