+++ RELEASE-NOTES.md
@@ -1,12 +1,17 @@
 ## RELEASE NOTES
 
-### Version 1.0.1 - December 31, 2018
+### Version 1.0.1 - January 1, 2019
 
 **NEW**
 - **Model** - Add `demo` attribute to video data model #1
 - **Feature** - Add demo lessons #2
+- **Model** - Create video collection data models #8
+- **Model** - Add `restricted_permit` attribute to device data model #9
+- **Feature** - Only permitted devices can access restricted videos #10
+- **Feature** - Add VB demo lessons, Y-GRE demo lessons, and Y-GRE test reviews #11
 
 **IMPROVED**
+- **Model** - Change `demo` attribute of video data model to `restricted` #7
 
 **FIXED**